fix strings3.rs
rposition requires a reversible iterator. It also needs a exact-size iterator,
so it is possible to assign indices the same way as position.
A exact-size iterator implements the std::iter::ExactSizeIterator trait,
but since not every iterator knows ahead of time how many items it will produce
(like .chars() which is used in the fv_char variable, which does not know AOT),
it cant be used on strings. Hence a byte array iterator is used instead, which
implements ExactSizeIterator.
And hence, is_ascii_whitespace is needed.